+++ src/tests/ast/res_msg_text.rs
@@ -43,6 +43,7 @@ fn typed() {
             id: Id::new_msg("hello-world"),
             comment: vec![],
             variables: vec![],
+            elements: vec![],
         }
     );
 }